requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.9
python-dotenv==1.0.1
requests==2.31.0
SQLAlchemy==2.0.28
typing_extensions==4.10.0
urllib3==2.0.3
Werkzeug==2.3.6
